refactor: 오목 패키지 재구성
