--- app/src/main/java/woowacourse/omok/Application.kt
@@ -1,13 +1,0 @@
-package woowacourse.omok
-
-import woowacourse.omok.controller.OmokController
-import woowacourse.omok.view.InputView
-import woowacourse.omok.view.OutputView
-
-fun main() {
-    val inputView = InputView()
-    val outputView = OutputView()
-
-    val controller = OmokController(inputView, outputView)
-    controller.run()
-}
