+++ app/src/main/java/woowacourse/omok/OmokApplication.kt
@@ -2,6 +2,9 @@ package woowacourse.omok
 
 import android.app.Application
 import woowacourse.omok.domain.repository.OmokRepository
+import woowacourse.omok.ui.controller.OmokController
+import woowacourse.omok.ui.view.InputView
+import woowacourse.omok.ui.view.OutputView
 
 class OmokApplication : Application() {
     lateinit var omokRepository: OmokRepository
@@ -11,4 +14,12 @@ class OmokApplication : Application() {
         super.onCreate()
         omokRepository = OmokRepository(applicationContext)
     }
+}
+
+fun main() {
+    val inputView = InputView()
+    val outputView = OutputView()
+
+    val controller = OmokController(inputView, outputView)
+    controller.run()
 }
