+++ app/src/main/java/woowacourse/omok/domain/model/omokboard/IntersectionState.kt
@@ -1,0 +1,7 @@
+package woowacourse.omok.domain.model.omokboard
+
+enum class IntersectionState {
+    OCCUPIED_BLACK,
+    OCCUPIED_WHITE,
+    EMPTY,
+}
