+++ app/src/main/java/woowacourse/omok/domain/model/omokboard/Position.kt
@@ -1,0 +1,6 @@
+package woowacourse.omok.domain.model.omokboard
+
+data class Position(
+    val row: Int,
+    val column: Int,
+)
