+++ app/src/main/java/woowacourse/omok/domain/model/player/PlayerStone.kt
@@ -1,0 +1,8 @@
+package woowacourse.omok.domain.model.player
+
+import woowacourse.omok.domain.model.omokboard.Position
+
+data class PlayerStone(
+    val color: StoneColor,
+    val position: Position,
+)
