+++ app/src/main/java/woowacourse/omok/domain/model/rule/OmokResult.kt
@@ -1,0 +1,3 @@
+package woowacourse.omok.domain.model.rule
+
+interface OmokResult
