+++ app/src/main/java/woowacourse/omok/domain/model/rule/judge/JudgeRule.kt
@@ -1,0 +1,5 @@
+package woowacourse.omok.domain.model.rule.judge
+
+import woowacourse.omok.domain.model.rule.OmokRule
+
+interface JudgeRule : OmokRule
