+++ app/src/main/java/woowacourse/omok/domain/model/rule/place/PlaceRule.kt
@@ -1,0 +1,5 @@
+package woowacourse.omok.domain.model.rule.place
+
+import woowacourse.omok.domain.model.rule.OmokRule
+
+interface PlaceRule : OmokRule
