--- app/src/main/java/woowacourse/omok/domain/omokboard/IntersectionState.kt
@@ -1,7 +1,0 @@
-package woowacourse.omok.domain.omokboard
-
-enum class IntersectionState {
-    OCCUPIED_BLACK,
-    OCCUPIED_WHITE,
-    EMPTY,
-}
