--- app/src/main/java/woowacourse/omok/domain/omokboard/Position.kt
@@ -1,6 +1,0 @@
-package woowacourse.omok.domain.omokboard
-
-data class Position(
-    val row: Int,
-    val column: Int,
-)
