--- app/src/main/java/woowacourse/omok/domain/player/StoneColor.kt
@@ -1,21 +1,0 @@
-package woowacourse.omok.domain.player
-
-import woowacourse.omok.domain.omokboard.IntersectionState
-
-enum class StoneColor {
-    BLACK,
-    WHITE,
-    ;
-
-    fun reversed(): StoneColor =
-        when (this) {
-            BLACK -> WHITE
-            WHITE -> BLACK
-        }
-
-    fun toIntersectionState(): IntersectionState =
-        when (this) {
-            BLACK -> IntersectionState.OCCUPIED_BLACK
-            WHITE -> IntersectionState.OCCUPIED_WHITE
-        }
-}
