--- app/src/main/java/woowacourse/omok/domain/rule/OmokResult.kt
@@ -1,3 +1,0 @@
-package woowacourse.omok.domain.rule
-
-interface OmokResult
