--- app/src/main/java/woowacourse/omok/domain/rule/OmokRule.kt
@@ -1,11 +1,0 @@
-package woowacourse.omok.domain.rule
-
-import woowacourse.omok.domain.omokboard.OmokBoard
-import woowacourse.omok.domain.player.PlayerStone
-
-interface OmokRule {
-    fun perform(
-        omokBoard: OmokBoard,
-        playerStone: PlayerStone,
-    ): OmokResult
-}
