--- app/src/main/java/woowacourse/omok/domain/rule/judge/JudgeResult.kt
@@ -1,16 +1,0 @@
-package woowacourse.omok.domain.rule.judge
-
-import woowacourse.omok.domain.player.StoneColor
-import woowacourse.omok.domain.rule.OmokResult
-
-sealed class JudgeResult : OmokResult {
-    sealed class Finished : JudgeResult() {
-        data class Win(
-            val stone: StoneColor,
-        ) : Finished()
-
-        data object Draw : Finished()
-    }
-
-    data object NotFinished : JudgeResult()
-}
