--- app/src/main/java/woowacourse/omok/domain/rule/judge/JudgeRule.kt
@@ -1,5 +1,0 @@
-package woowacourse.omok.domain.rule.judge
-
-import woowacourse.omok.domain.rule.OmokRule
-
-interface JudgeRule : OmokRule
