--- app/src/main/java/woowacourse/omok/domain/rule/place/AlreadyExistStoneRule.kt
@@ -1,17 +1,0 @@
-package woowacourse.omok.domain.rule.place
-
-import woowacourse.omok.domain.omokboard.IntersectionState
-import woowacourse.omok.domain.omokboard.OmokBoard
-import woowacourse.omok.domain.player.PlayerStone
-
-class AlreadyExistStoneRule : PlaceRule {
-    override fun perform(
-        omokBoard: OmokBoard,
-        playerStone: PlayerStone,
-    ): PlaceResult =
-        if (omokBoard.find(playerStone.position) == IntersectionState.EMPTY) {
-            PlaceResult.Success
-        } else {
-            PlaceResult.Failure.AlreadyExistStone
-        }
-}
