--- app/src/main/java/woowacourse/omok/domain/rule/place/PlaceRule.kt
@@ -1,5 +1,0 @@
-package woowacourse.omok.domain.rule.place
-
-import woowacourse.omok.domain.rule.OmokRule
-
-interface PlaceRule : OmokRule
